Total reorganization
This was one of my first nodejs projects and I made a real mess of the organizational aspect. It was just a hodgepodge of techniques that were ctrl-v from all over the place. This seperates one into 	`server` and `client` directories and tries to make sense of a mad mad world.